@@ -83,31 +82,31 @@
     "ember-modal-dialog": "^4.1.2",
     "ember-page-title": "^8.0.0",
     "ember-popper-modifier": "^3.0.0",
-    "ember-power-select": "^7.1.1",
+    "ember-power-select": "^7.1.2",
     "ember-print-this": "^2.0.0",
-    "ember-qunit": "^7.0.0",
+    "ember-qunit": "^8.0.1",
     "ember-resolver": "^11.0.1",
     "ember-source": "~4.12.0",
     "ember-style-modifier": "^3.0.1",
-    "ember-svg-jar": "^2.4.4",
+    "ember-svg-jar": "^2.4.6",
     "ember-template-lint": "^5.11.2",
     "ember-try": "^3.0.0",
     "ember-web-workers": "^1.0.0",
     "ember-welcome-page": "^7.0.2",
-    "eslint": "^8.47.0",
+    "eslint": "^8.51.0",
     "eslint-config-prettier": "^9.0.0",
-    "eslint-plugin-ember": "^11.11.0",
+    "eslint-plugin-ember": "^11.11.1",
     "eslint-plugin-import": "^2.28.1",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-prettier": "^5.0.0",
-    "eslint-plugin-qunit": "^8.0.0",
+    "eslint-plugin-prettier": "^5.0.1",
+    "eslint-plugin-qunit": "^8.0.1",
     "loader.js": "^4.7.0",
     "npm-run-all": "^4.1.5",
-    "prettier": "^3.0.2",
-    "qunit": "^2.19.4",
-    "qunit-dom": "^2.0.0",
-    "typescript": "5.1.6",
-    "webpack": "^5.88.2",
+    "prettier": "^3.0.3",
+    "qunit": "^2.20.0",
+    "qunit-dom": "^3.0.0",
+    "typescript": "5.2.2",
+    "webpack": "^5.89.0",
     "yuidoc-bootstrap-theme": "^1.0.6",
     "yuidocjs": "^0.10.2"
   },
