@@ -1,4 +1,7 @@
-import { getAllClassesInApplication } from './application-helpers';
+import {
+  getAllClassesInApplication,
+  getAllPackagesInApplication,
+} from './application-helpers';
 import {
   StructureLandscapeData,
   Node,
@@ -7,19 +10,56 @@ import {
   Class,
   isPackage,
   Method,
+  isApplication,
 } from './landscape-schemes/structure-data';
 import { getApplicationFromPackage } from './landscape-structure-helpers';
 import sha256 from 'crypto-js/sha256';
 import { DrawableClassCommunication } from './application-rendering/class-communication-computer';
-import { getAncestorPackages, getClassesInPackage } from './package-helpers';
+import {
+  getAncestorPackages,
+  getClassesInPackage,
+  getSubPackagesOfPackage,
+} from './package-helpers';
 
-export function setClassName(app: Application, id: string) {
+export enum EntityType {
+  App = 'APP',
+  Package = 'PACKAGE',
+  SubPackage = 'SUBPACKAGE',
+  Clazz = 'CLAZZ',
+  Communication = 'COMMUNICATION',
+}
+
+export enum RestructureAction {
+  Create = 'CREATE',
+  Rename = 'RENAME',
+  Delete = 'DELETE',
+  CopyPaste = 'COPYPASTE',
+  CutInsert = 'CUTINSERT',
+  Communication = 'COMMUNICATION',
+}
+
+/**
+ * Finds a class from an application based on the class ID.
+ *
+ * @param app The target application to search within.
+ * @param id The ID of the class to retrieve.
+ * @returns The class matching the given ID, or undefined if not found.
+ */
+export function getClassInApplicationById(app: Application, id: string) {
   const allClassesInApplication = getAllClassesInApplication(app);
   const classToRename = allClassesInApplication.find((cls) => cls.id === id);
 
   return classToRename;
 }
 
+/**
+ * Adds a new node with an application to the landscape
+ *
+ * @param appName The name of the new application.
+ * @param language The programming language of the application.
+ * @param counter A unique identifier to aid in creating unique entities.
+ * @returns The newly created node with its child entities.
+ */
 export function addFoundationToLandscape(
   appName: string,
   language: string,
@@ -41,9 +81,12 @@ export function addFoundationToLandscape(
     packages: [],
   };
 
-  const newPckg = createPackage(counter);
+  const newPckg = createPackage(
+    'newPackage' + counter,
+    'New Package ' + counter
+  );
 
-  const newClass = createClass(counter);
+  const newClass = createClass('newClass' + counter, 'New Class ' + counter);
   newClass.parent = newPckg;
 
   newPckg.classes.push(newClass as Class);
@@ -54,10 +97,64 @@ export function addFoundationToLandscape(
   return myNode;
 }
 
-export function createPackage(counter: number) {
+export function duplicateApplication(
+  landscapeData: StructureLandscapeData,
+  app: Application,
+  wrapper: {
+    idCounter: number;
+    comms: DrawableClassCommunication[];
+    copiedComms: DrawableClassCommunication[];
+  }
+) {
+  const originalNode = app.parent;
+  const duplicatedNode: Node = {
+    id: originalNode.id,
+    ipAddress: originalNode.ipAddress,
+    hostName: originalNode.hostName,
+    applications: [],
+  };
+
+  const duplicatedApp: Application = {
+    id: app.id,
+    name: app.name,
+    language: app.language,
+    instanceId: app.instanceId,
+    parent: duplicatedNode,
+    packages: [],
+  };
+
+  duplicatedNode.applications.pushObject(duplicatedApp);
+
+  app.packages.forEach((pckg) => {
+    const duplicatedPackage = copyPackageContent(pckg);
+    duplicatedApp.packages.pushObject(duplicatedPackage);
+  });
+
+  // Copy the communications
+  if (wrapper.comms.length > 0) {
+    const classesInApp = getAllClassesInApplication(duplicatedApp);
+
+    copyCommunications(classesInApp, wrapper, duplicatedApp);
+  }
+
+  changeID({ entity: duplicatedApp }, 'duplicated|');
+
+  landscapeData.nodes.pushObject(duplicatedNode);
+
+  return duplicatedApp;
+}
+
+/**
+ * Creates a new package with the given ID and name.
+ *
+ * @param id The unique identifier for the package.
+ * @param name The name of the package.
+ * @returns A new package object.
+ */
+export function createPackage(id: string, name: string) {
   const newPckg: Package = {
-    id: 'newPackage ' + counter,
-    name: 'New Package ' + counter,
+    id: id,
+    name: name,
     subPackages: [],
     classes: [],
   };
@@ -65,344 +162,331 @@ export function createPackage(counter: number) {
   return newPckg;
 }
 
-export function createClass(counter: number) {
+/**
+ * Creates a new class with the given ID and name.
+ *
+ * @param id The unique identifier for the class.
+ * @param name The name of the class.
+ * @returns A new class object.
+ */
+export function createClass(id: string, name: string) {
   const newClass: Partial<Class> = {
-    id: 'newClass ' + counter,
-    name: 'New Class ' + counter,
+    id: id,
+    name: name,
     methods: [],
   };
 
   return newClass;
 }
 
-export function addMethodToClass(clazz: Class, methodName: string) {
-  const newMethod: Method = {
-    name: methodName,
-    hashCode: sha256(methodName).toString(),
-  };
+/**
+ * Adds a method to a class. If the method is a copy, the hash code will simply be the method name itself.
+ *
+ * @param clazz The class to which the method will be added.
+ * @param methodName The name of the method.
+ * @param isCopy A flag indicating if the method is a copy.
+ */
+export function addMethodToClass(
+  clazz: Class,
+  methodName: string,
+  isCopy: boolean = false
+) {
+  const newMethod: Method = !isCopy
+    ? {
+        name: methodName,
+        hashCode: sha256(methodName).toString(),
+      }
+    : {
+        name: methodName,
+        hashCode: methodName,
+      };
+
   clazz.methods.push(newMethod);
 }
 
-export function removeApplication(
-  landscapeStructure: StructureLandscapeData,
-  wrapper: {
-    comms: DrawableClassCommunication[];
-    meshTodelete?: Application | Package | Class;
-  },
-  app: Application,
-  isCutOperation: boolean,
-  destinationApplication?: Application
-) {
-  if (wrapper.comms.length > 0) {
-    const classesInApplication = getAllClassesInApplication(app);
-    if (isCutOperation) {
-      // if any class in application is part of a communication and information about source and target app has changed then update it
-      updateAffectedCommunications(
-        classesInApplication,
-        wrapper,
-        destinationApplication
-      );
-    } else {
-      // if any class in application is part of a communication then remove it
-      removeAffectedCommunications(classesInApplication, wrapper);
-    }
+/**
+ * Copies package with its content.
+ * @param pckgToCopy The package to be copied
+ * @returns A new package with copied content
+ */
+export function copyPackageContent(pckgToCopy: Package): Package {
+  // Create a new package with the same ID and name
+  const copiedPckg = createPackage(pckgToCopy.id, pckgToCopy.name);
+
+  // Copy each subpackage and then add to the copied class
+  for (const subPckg of pckgToCopy.subPackages) {
+    const copiedSubPckg = copyPackageContent(subPckg);
+    copiedSubPckg.parent = copiedPckg;
+    copiedPckg.subPackages.push(copiedSubPckg);
   }
 
-  if (isCutOperation) {
-    const parentNode = app.parent;
-    // If the node contains only of the one application, then delete the whole node, otherwise remove application from node.
-    if (parentNode.applications.length <= 1) {
-      landscapeStructure.nodes = landscapeStructure.nodes.filter(
-        (node) => node.id != parentNode.id
-      );
-    } else {
-      parentNode.applications = parentNode.applications.filter(
-        (appl) => appl.id != app.id
-      );
-    }
+  // Copy each class and then add to the copied package
+  for (const clazz of pckgToCopy.classes) {
+    const copiedClass = copyClassContent(clazz);
+    copiedClass.parent = copiedPckg;
+    copiedPckg.classes.push(copiedClass);
   }
 
-  wrapper.meshTodelete = app;
-  // if (commsWrapper.app) {
-  //   commsWrapper.app = app;
-  // }
-
-  //app.packages = [];
+  return copiedPckg;
 }
 
-export function removePackageFromApplication(
+/**
+ * Copies a class
+ * @param clazzToCopy The Class to be copied
+ * @returns A new class with with copied content
+ */
+export function copyClassContent(clazzToCopy: Class) {
+  // Create a new class with the same ID and name
+  const copiedClass = createClass(clazzToCopy.id, clazzToCopy.name) as Class;
+
+  // Copy each method and add to the copied class
+  for (const method of clazzToCopy.methods) {
+    addMethodToClass(copiedClass, method.name, true);
+  }
+
+  return copiedClass;
+}
+
+/**
+ * Removes an application in the landscape
+ * @param landscapeStructure The landscape data which contains the app
+ * @param app The application that needs to be removed
+ */
+export function removeApplication(
   landscapeStructure: StructureLandscapeData,
-  wrapper: {
-    comms: DrawableClassCommunication[];
-    meshTodelete?: Application | Package | Class;
-  },
+  app: Application
+) {
+  const parentNode = app.parent;
+
+  //If the node contains only of the one application, then delete the whole node, otherwise remove application from node.
+  if (parentNode.applications.length <= 1) {
+    landscapeStructure.nodes = landscapeStructure.nodes.filter(
+      (node) => node.id != parentNode.id
+    );
+  } else {
+    parentNode.applications = parentNode.applications.filter(
+      (appl) => appl.id != app.id
+    );
+  }
+}
+
+/**
+ * Removes a package within an application
+ * @param pckgToRemove The package that needs to be removed
+ * @param app The application where the package resides in
+ */
+export function removePackageFromApplication(
   pckgToRemove: Package,
-  isCutOperation: boolean,
-  destinationApplication?: Application
+  app: Application
 ) {
   const parentPackage = pckgToRemove.parent;
 
-  // if parent is another package then remove current package from parents subpackage list, else remove the whole application
   if (parentPackage) {
-    const applicationWrapper: { app: Application | undefined } = {
-      app: undefined,
-    };
-    let deleteApp = false;
-
-    // if parent package has more than 1 element, then remove the current package from its child packages, else check for the next ancestors
-    if (parentPackage.subPackages.length + parentPackage.classes.length > 1) {
-      wrapper.meshTodelete = pckgToRemove;
-      if (isCutOperation) {
-        parentPackage.subPackages = parentPackage.subPackages.filter(
-          (packg) => packg.id != pckgToRemove.id
-        );
-      }
-      //delete pckgToRemove.parent;
-    } else {
-      // check if there is one ancestor package, that fulfills the condition of having more than 1 child elements. If not then we delete the whole application
-      deleteApp = cleanUpAncestor(
-        landscapeStructure,
-        parentPackage,
-        applicationWrapper,
-        wrapper,
-        isCutOperation
-      );
-    }
-
-    if (deleteApp && applicationWrapper.app) {
-      removeApplication(
-        landscapeStructure,
-        wrapper,
-        applicationWrapper.app,
-        isCutOperation,
-        destinationApplication
-      );
-    } else if (!deleteApp && wrapper.comms.length > 0) {
-      const classesInPackage = getClassesInPackage(pckgToRemove);
-      if (isCutOperation) {
-        // if any class in package is part of a communication and information about source and target app has changed then update it
-        updateAffectedCommunications(
-          classesInPackage,
-          wrapper,
-          destinationApplication
-        );
-      } else {
-        // if any class in package is part of a communication then remove it
-        removeAffectedCommunications(classesInPackage, wrapper);
-      }
-    }
-  } else {
-    const parentApplication = getApplicationFromPackage(
-      landscapeStructure,
-      pckgToRemove.id
+    parentPackage.subPackages = parentPackage.subPackages.filter(
+      (packg) => packg.id != pckgToRemove.id
     );
-
-    // if applications only package is this package, then remove whole application, else remove package from applications package list
-    if (parentApplication && parentApplication.packages.length <= 1) {
-      removeApplication(
-        landscapeStructure,
-        wrapper,
-        parentApplication,
-        isCutOperation,
-        destinationApplication
-      );
-    } else if (parentApplication && parentApplication.packages.length > 1) {
-      wrapper.meshTodelete = pckgToRemove;
-      if (isCutOperation) {
-        parentApplication.packages = parentApplication.packages.filter(
-          (pckg) => pckg.id !== pckgToRemove.id
-        );
-      }
-
-      if (wrapper.comms.length > 0) {
-        // if any class in application is part of a communication then remove it
-        const classesInPackage = getClassesInPackage(pckgToRemove);
-        if (isCutOperation) {
-          // if any class in package is part of a communication and information about source and target app has changed then update it
-          updateAffectedCommunications(
-            classesInPackage,
-            wrapper,
-            destinationApplication
-          );
-        } else {
-          // if any class in package is part of a communication then remove it
-          removeAffectedCommunications(classesInPackage, wrapper);
-        }
-      }
-    }
+  } else {
+    app.packages = app.packages.filter((pckg) => pckg.id !== pckgToRemove.id);
   }
 }
 
-export function removeClassFromPackage(
-  landscapeStructure: StructureLandscapeData,
-  wrapper: {
-    comms: DrawableClassCommunication[];
-    meshTodelete?: Application | Package | Class;
-  },
-  clazzToRemove: Class,
-  isCutOperation: boolean,
-  destinationApplication?: Application
-) {
+/**
+ * Determines whether a package can be removed or not
+ * @param pckgToRemove The package to approve the removal
+ * @param app The application where the package resides in
+ * @returns Approve of removal
+ */
+export function canDeletePackage(pckgToRemove: Package, app: Application) {
+  const parentPackage = pckgToRemove.parent;
+  if (parentPackage) {
+    return parentPackage.subPackages.length + parentPackage.classes.length > 1;
+  } else {
+    return app && app.packages.length > 1;
+  }
+}
+
+/**
+ * Removes a class within a package.
+ * @param clazzToRemove The class that needs to be removed
+ */
+export function removeClassFromPackage(clazzToRemove: Class) {
   const parentPackage = clazzToRemove.parent;
   if (parentPackage) {
-    // if parent Package has more than 1 element, then remove the class from it else remove the parent Package
-    if (parentPackage.subPackages.length + parentPackage.classes.length > 1) {
-      wrapper.meshTodelete = clazzToRemove;
-      if (isCutOperation) {
-        parentPackage.classes = parentPackage.classes.filter(
-          (clzz) => clzz.id != clazzToRemove.id
-        );
-      }
-      if (wrapper.comms.length > 0) {
-        if (isCutOperation) {
-          // if class in package is part of a communication and information about source and target app has changed then update it
-          wrapper.comms.forEach((comms) => {
-            if (comms.sourceClass.id === clazzToRemove.id)
-              comms.sourceApp = destinationApplication;
-            else if (comms.targetClass.id === clazzToRemove.id)
-              comms.targetApp = destinationApplication;
-          });
-        } else {
-          // if class in package is part of a communication then remove it
-          wrapper.comms = wrapper.comms.filter(
-            (comm) =>
-              comm.sourceClass.id !== clazzToRemove.id &&
-              comm.targetClass.id !== clazzToRemove.id
-          );
-        }
-      }
-    } else {
-      removePackageFromApplication(
-        landscapeStructure,
-        wrapper,
-        parentPackage,
-        isCutOperation,
-        destinationApplication
-      );
-    }
+    parentPackage.classes = parentPackage.classes.filter(
+      (clzz) => clzz.id != clazzToRemove.id
+    );
   }
 }
 
-export function cutAndInsertPackage(
+/**
+ * Determines whether a class can be removed or not
+ * @param clazzToRemove The class to approve of the removal
+ * @returns Approval of removal
+ */
+export function canDeleteClass(clazzToRemove: Class) {
+  const parentPackage = clazzToRemove.parent;
+  return parentPackage.classes.length + parentPackage.subPackages.length > 1;
+}
+
+// TODO this and cutAndInsertPackage have almost identical code! Needs to be refactored!!!!
+export function pastePackage(
+  landscapeStructure: StructureLandscapeData,
+  copiedPackage: Package,
+  pasteToDestination: Application | Package,
+  wrapper: {
+    idCounter: number;
+    comms: DrawableClassCommunication[];
+    copiedComms: DrawableClassCommunication[];
+  }
+) {
+  let destinationApplication: Application | undefined;
+
+  if (isPackage(pasteToDestination)) {
+    // Get the main application containing the destimination package
+    const firstPackage = getAncestorPackages(pasteToDestination);
+    if (firstPackage.length > 0)
+      destinationApplication = getApplicationFromPackage(
+        landscapeStructure,
+        firstPackage[firstPackage.length - 1].id
+      );
+    else
+      destinationApplication = getApplicationFromPackage(
+        landscapeStructure,
+        pasteToDestination.id
+      );
+
+    // Insert the package to be moved under the destination package
+    insertPackageToPackage(pasteToDestination, copiedPackage);
+  } else {
+    // If the destination is an application, insert the package directly under it
+    insertPackageToApplication(pasteToDestination, copiedPackage);
+    destinationApplication = pasteToDestination;
+  }
+
+  // Copy the communications
+  if (wrapper.comms.length > 0) {
+    const classesInPackage = getClassesInPackage(copiedPackage);
+
+    copyCommunications(classesInPackage, wrapper, destinationApplication);
+  }
+}
+
+export function pasteClass(
+  landscapeStructure: StructureLandscapeData,
+  copiedClass: Class,
+  pasteToDestination: Package,
+  commsWrapper: {
+    idCounter: number;
+    comms: DrawableClassCommunication[];
+    copiedComms: DrawableClassCommunication[];
+  }
+) {
+  // Verify if the destination is a package
+  if (isPackage(pasteToDestination)) {
+    let destinationApplication: Application | undefined;
+    const firstPackage = getAncestorPackages(pasteToDestination);
+
+    // Find the application of destination package
+    if (firstPackage.length > 0)
+      destinationApplication = getApplicationFromPackage(
+        landscapeStructure,
+        firstPackage[firstPackage.length - 1].id
+      );
+    else
+      destinationApplication = getApplicationFromPackage(
+        landscapeStructure,
+        pasteToDestination.id
+      );
+
+    if (commsWrapper.comms.length) {
+      copyCommunications([copiedClass], commsWrapper, destinationApplication);
+    }
+
+    // Add the copied class to the destination package's classes
+    pasteToDestination.classes.push(copiedClass);
+    copiedClass.parent = pasteToDestination;
+  }
+}
+
+/**
+ * Cuts a specified package from its current location in the landscape structure and inserts it into a given destination,
+ * which can be another package or an application. The class communications are appropriately updated.
+ * @param landscapeStructure The data structure representing the current state of the landscape.
+ * @param clippedPackage The package that is intended to be moved.
+ * @param destination The destination where the package should be inserted. This can be either an application or another package.
+ * @param wrapper Contains data on drawable class communications and an optional mesh object that might need deletion
+ *                after the package move operation.
+ */
+export function movePackage(
   landscapeStructure: StructureLandscapeData,
   clippedPackage: Package,
-  clipToDestination: Application | Package,
+  destination: Application | Package,
   wrapper: {
     comms: DrawableClassCommunication[];
     meshTodelete?: Application | Package | Class;
+    updatedComms: DrawableClassCommunication[];
   }
 ) {
-  const parentPackage = clippedPackage.parent;
+  let destinationApplication: Application | undefined;
 
-  if (parentPackage) {
-    let destinationApplication: Application | undefined;
-    // if the destination is a package, then push the clipped object in its subpackage, else push it in the applications packages
-    if (isPackage(clipToDestination)) {
-      const firstPackage = getAncestorPackages(clipToDestination);
-      if (firstPackage.length > 0)
-        destinationApplication = getApplicationFromPackage(
-          landscapeStructure,
-          firstPackage[firstPackage.length - 1].id
-        );
-      else
-        destinationApplication = getApplicationFromPackage(
-          landscapeStructure,
-          clipToDestination.id
-        );
-      insertClipToDestinationPackage(clipToDestination, clippedPackage);
-    } else {
-      inserClipToDestinationApp(clipToDestination, clippedPackage);
-      destinationApplication = clipToDestination;
-    }
-    if (parentPackage.subPackages.length + parentPackage.classes.length <= 1) {
-      removePackageFromApplication(
+  if (isPackage(destination)) {
+    // Get the main application containing the destimination package
+    const firstPackage = getAncestorPackages(destination);
+    if (firstPackage.length > 0)
+      destinationApplication = getApplicationFromPackage(
         landscapeStructure,
-        wrapper,
-        parentPackage,
-        true,
-        destinationApplication
+        firstPackage[firstPackage.length - 1].id
       );
-    } else {
-      parentPackage.subPackages = parentPackage.subPackages.filter(
-        (packg: Package) => packg.id != clippedPackage.id
+    else
+      destinationApplication = getApplicationFromPackage(
+        landscapeStructure,
+        destination.id
       );
-      wrapper.meshTodelete = clippedPackage;
-      if (wrapper.comms.length > 0) {
-        const classesInPackage = getClassesInPackage(clippedPackage);
-        updateAffectedCommunications(
-          classesInPackage,
-          wrapper,
-          destinationApplication
-        );
-      }
-    }
+
+    // Insert the package to be moved under the destination package
+    insertPackageToPackage(destination, clippedPackage);
   } else {
-    const application = getApplicationFromPackage(
-      landscapeStructure,
-      clippedPackage.id
+    // If the destination is an application, insert the package directly under it
+    insertPackageToApplication(destination, clippedPackage);
+    destinationApplication = destination;
+  }
+
+  // Update communications if the package contains classes involved in communications
+  if (wrapper.comms.length > 0) {
+    const classesInPackage = getClassesInPackage(clippedPackage);
+
+    updateAffectedCommunications(
+      classesInPackage,
+      wrapper,
+      destinationApplication
     );
-    let destinationApplication: Application | undefined;
-
-    // if the destination is a package, then push the clipped object in its subpackage, else push it in the applications packages
-    if (isPackage(clipToDestination)) {
-      const firstPackage = getAncestorPackages(clipToDestination);
-      if (firstPackage.length > 0)
-        destinationApplication = getApplicationFromPackage(
-          landscapeStructure,
-          firstPackage[firstPackage.length - 1].id
-        );
-      else
-        destinationApplication = getApplicationFromPackage(
-          landscapeStructure,
-          clipToDestination.id
-        );
-      insertClipToDestinationPackage(clipToDestination, clippedPackage);
-    } else {
-      destinationApplication = clipToDestination;
-      inserClipToDestinationApp(clipToDestination, clippedPackage);
-    }
-
-    if (application && application.packages.length <= 1) {
-      removeApplication(
-        landscapeStructure,
-        wrapper,
-        application,
-        true,
-        destinationApplication
-      );
-    } else if (application && application.packages.length > 1) {
-      // remove clipped object from parent application packages
-      application.packages = application.packages.filter(
-        (packg) => packg.id !== clippedPackage.id
-      );
-      wrapper.meshTodelete = clippedPackage;
-
-      // if any class in package is part of a communication and information about source and target app has changed then update it
-      if (wrapper.comms.length > 0) {
-        const classesInPackage = getClassesInPackage(clippedPackage);
-        updateAffectedCommunications(
-          classesInPackage,
-          wrapper,
-          destinationApplication
-        );
-      }
-    }
   }
 }
 
-export function cutAndInsertClass(
+/**
+ * Moves a class from its current package and inserts it into a selected package.
+ *
+ * @param landscapeStructure The data structure representing the current state of the landscape.
+ * @param clippedClass The class to be moved.
+ * @param clipToDestination The destination package where the class will be moved to.
+ * @param commsWrapper Contains an array of class communications.
+ */
+export function moveClass(
   landscapeStructure: StructureLandscapeData,
   clippedClass: Class,
   clipToDestination: Package,
-  commsWrapper: { comms: DrawableClassCommunication[] }
+  commsWrapper: {
+    comms: DrawableClassCommunication[];
+    meshTodelete?: Application | Package | Class;
+    updatedComms?: DrawableClassCommunication[];
+  }
 ) {
-  const parentPackage = clippedClass.parent;
-
+  // Verify if the destination is a package
   if (isPackage(clipToDestination)) {
     let destinationApplication: Application | undefined;
     const firstPackage = getAncestorPackages(clipToDestination);
 
-    // find application of destination
+    // Find the application of destination package
     if (firstPackage.length > 0)
       destinationApplication = getApplicationFromPackage(
         landscapeStructure,
@@ -414,243 +498,225 @@ export function cutAndInsertClass(
         clipToDestination.id
       );
 
-    if (parentPackage.subPackages.length + parentPackage.classes.length <= 1) {
-      removePackageFromApplication(
-        landscapeStructure,
+    if (commsWrapper.comms.length) {
+      updateAffectedCommunications(
+        [clippedClass],
         commsWrapper,
-        parentPackage,
-        true,
-        destinationApplication
-      );
-    } else {
-      removeClassFromPackage(
-        landscapeStructure,
-        commsWrapper,
-        clippedClass,
-        true,
         destinationApplication
       );
     }
 
+    // Add the moved class to the destination package's classes
     clipToDestination.classes.push(clippedClass);
     clippedClass.parent = clipToDestination;
   }
 }
 
+export function changeID(
+  wrapper: { entity: Application | Package | Class },
+  id: string
+) {
+  if (isApplication(wrapper.entity)) {
+    const node = wrapper.entity.parent;
+    node.id = id + node.id;
+
+    wrapper.entity.id = id + wrapper.entity.id;
+
+    const allPackages = getAllPackagesInApplication(wrapper.entity);
+    const allClasses = getAllClassesInApplication(wrapper.entity);
+
+    allPackages.forEach((pckg) => {
+      pckg.id = id + pckg.id;
+    });
+
+    allClasses.forEach((clazz) => {
+      clazz.id = id + clazz.id;
+    });
+  } else if (isPackage(wrapper.entity)) {
+    const allSubPackages = getSubPackagesOfPackage(wrapper.entity, true);
+    const allClassesInPackge = getClassesInPackage(wrapper.entity, true);
+
+    wrapper.entity.id = id + wrapper.entity.id;
+
+    allSubPackages.forEach((pckg) => {
+      pckg.id = id + pckg.id;
+    });
+
+    allClassesInPackge.forEach((clazz) => {
+      clazz.id = id + clazz.id;
+      if (clazz.methods.length) {
+        clazz.methods.forEach((method) => {
+          method.hashCode = id + method.hashCode;
+        });
+      }
+    });
+  } else {
+    wrapper.entity.id = id + wrapper.entity.id;
+    if (wrapper.entity.methods.length) {
+      wrapper.entity.methods.forEach((method) => {
+        method.hashCode = id + method.hashCode;
+      });
+    }
+  }
+}
+
+export function restoreID(
+  wrapper: { entity: Application | Package | Class },
+  prependID: string
+) {
+  if (isApplication(wrapper.entity)) {
+    const allPackages = getAllPackagesInApplication(wrapper.entity);
+    const allClasses = getAllClassesInApplication(wrapper.entity);
+
+    allPackages.forEach((pckg) => {
+      pckg.id = removePrependFromID(pckg.id, prependID);
+    });
+
+    allClasses.forEach((clazz) => {
+      clazz.id = removePrependFromID(clazz.id, prependID);
+    });
+  } else if (isPackage(wrapper.entity)) {
+    const allSubPackages = getSubPackagesOfPackage(wrapper.entity, true);
+    const allClassesInPackage = getClassesInPackage(wrapper.entity, true);
+
+    wrapper.entity.id = removePrependFromID(wrapper.entity.id, prependID);
+
+    allSubPackages.forEach((pckg) => {
+      pckg.id = removePrependFromID(pckg.id, prependID);
+    });
+
+    allClassesInPackage.forEach((clazz) => {
+      clazz.id = removePrependFromID(clazz.id, prependID);
+      if (clazz.methods.length) {
+        clazz.methods.forEach((method) => {
+          method.hashCode = removePrependFromID(method.hashCode, prependID);
+        });
+      }
+    });
+  } else {
+    wrapper.entity.id = removePrependFromID(wrapper.entity.id, prependID);
+    if (wrapper.entity.methods.length) {
+      wrapper.entity.methods.forEach((method) => {
+        method.hashCode = removePrependFromID(method.hashCode, prependID);
+      });
+    }
+  }
+}
+
+function removePrependFromID(changedID: string, prepend: string) {
+  if (changedID.startsWith(prepend)) {
+    return changedID.substring(prepend.length);
+  }
+  return changedID;
+}
+
 function updateAffectedCommunications(
   classesInPackage: Class[],
-  commsWrapper: { comms: DrawableClassCommunication[] },
+  commsWrapper: {
+    comms: DrawableClassCommunication[];
+    updatedComms?: DrawableClassCommunication[];
+  },
   destinationApplication: Application | undefined
 ) {
   classesInPackage.forEach((clazz) =>
     commsWrapper.comms.forEach((comms) => {
-      if (comms.sourceClass.id === clazz.id)
+      if (comms.sourceClass.id === clazz.id) {
+        comms.sourceClass = clazz;
         comms.sourceApp = destinationApplication;
-      else if (comms.targetClass.id === clazz.id)
+        commsWrapper.updatedComms?.push(comms);
+      } else if (comms.targetClass.id === clazz.id) {
+        comms.targetClass = clazz;
         comms.targetApp = destinationApplication;
+        commsWrapper.updatedComms?.push(comms);
+      }
     })
   );
 }
 
-function inserClipToDestinationApp(
-  clipToDestination: Application,
+function copyCommunications(
+  classesInPackage: Class[],
+  commsWrapper: {
+    idCounter: number;
+    comms: DrawableClassCommunication[];
+    copiedComms: DrawableClassCommunication[];
+  },
+  destinationApplication: Application | undefined
+) {
+  const copiedCommsMap = new Map();
+  classesInPackage.forEach((clazz) =>
+    commsWrapper.comms.forEach((comms) => {
+      if (
+        comms.sourceClass.id === clazz.id ||
+        comms.targetClass.id === clazz.id
+      ) {
+        let copiedComms;
+        // If this communication has already been copied, retrieve the copy from the map, otherwise create a new copy and store it in the map
+        if (copiedCommsMap.has(comms.id)) {
+          copiedComms = copiedCommsMap.get(comms.id);
+        } else {
+          copiedComms = copyCommunication(comms, commsWrapper.idCounter);
+          copiedCommsMap.set(comms.id, copiedComms);
+        }
+
+        if (comms.sourceClass.id === clazz.id) {
+          copiedComms.sourceClass = clazz;
+          copiedComms.sourceApp = destinationApplication;
+        }
+        if (comms.targetClass.id === clazz.id) {
+          copiedComms.targetClass = clazz;
+          copiedComms.targetApp = destinationApplication;
+        }
+      }
+    })
+  );
+  commsWrapper.copiedComms = Array.from(copiedCommsMap.values());
+}
+
+function copyCommunication(
+  commToCopy: DrawableClassCommunication,
+  idCounter: number
+) {
+  const comm: DrawableClassCommunication = {
+    id: 'copied' + idCounter + '|' + commToCopy.id,
+    totalRequests: commToCopy.totalRequests,
+    sourceClass: commToCopy.sourceClass,
+    targetClass: commToCopy.targetClass,
+    operationName: 'copied' + idCounter + '|' + commToCopy.operationName,
+    sourceApp: commToCopy.sourceApp,
+    targetApp: commToCopy.targetApp,
+  };
+
+  return comm;
+}
+
+function insertPackageToApplication(
+  destinationApp: Application,
   clippedPackage: Package
 ) {
-  clipToDestination.packages.push(clippedPackage);
+  destinationApp.packages.push(clippedPackage);
   delete clippedPackage.parent;
 }
 
-function insertClipToDestinationPackage(
-  clipToDestination: Package,
+function insertPackageToPackage(
+  destinationPckg: Package,
   clippedPackage: Package
 ) {
-  clipToDestination.subPackages.push(clippedPackage);
-  clippedPackage.parent = clipToDestination;
+  destinationPckg.subPackages.push(clippedPackage);
+  clippedPackage.parent = destinationPckg;
 }
 
-function removeAffectedCommunications(
+export function removeAffectedCommunications(
   classesInApplication: Class[],
-  commsWrapper: { comms: DrawableClassCommunication[] }
-) {
-  classesInApplication.forEach(
-    (clazz) =>
-      (commsWrapper.comms = commsWrapper.comms.filter(
-        (comm) =>
-          comm.sourceClass.id !== clazz.id && comm.targetClass.id !== clazz.id
-      ))
-  );
-}
-
-function cleanUpAncestor(
-  landscapeStructure: StructureLandscapeData,
-  parentPackage: Package,
-  applicationWrapper: { app: Application | undefined },
-  wrapper: {
+  commsWrapper: {
     comms: DrawableClassCommunication[];
-    meshTodelete?: Application | Package | Class;
-  },
-  isCutOperation?: boolean
-) {
-  const ancestorPackages = getAncestorPackages(parentPackage);
-  /*  case 1: ancestor is the top Package meaning the highest on the hierarchy.
-      case 2: there are multiple ancestors. We need to find the nearest ancestor with more than 1 element in it
-      case 3: there are no ancestors since the current package is top Package, so we investigate the application
-  */
-  if (ancestorPackages.length == 1)
-    return handleTopPackageAncestor(
-      landscapeStructure,
-      parentPackage,
-      ancestorPackages,
-      applicationWrapper,
-      wrapper,
-      isCutOperation
-    );
-  else if (ancestorPackages.length > 1)
-    return handlePackageAncestors(
-      landscapeStructure,
-      parentPackage,
-      ancestorPackages,
-      applicationWrapper,
-      wrapper,
-      isCutOperation
-    );
-  else
-    return handleApplicationAncestor(
-      landscapeStructure,
-      parentPackage,
-      applicationWrapper,
-      wrapper,
-      isCutOperation
-    );
-}
-
-function handleApplicationAncestor(
-  landscapeStructure: StructureLandscapeData,
-  parentPackage: Package,
-  applicationWrapper: { app: Application | undefined },
-  wrapper: {
-    comms: DrawableClassCommunication[];
-    meshTodelete?: Application | Package | Class;
-  },
-  isCutOperation?: boolean
-) {
-  //check if the whole application has still more than 1 packages and if so then remove only the top package else the whole application
-  applicationWrapper.app = getApplicationFromPackage(
-    landscapeStructure,
-    parentPackage.id
-  );
-  return hasAppEnoughPackages(
-    applicationWrapper,
-    parentPackage,
-    wrapper,
-    isCutOperation
-  );
-}
-
-function handlePackageAncestors(
-  landscapeStructure: StructureLandscapeData,
-  parentPackage: Package,
-  ancestorPackages: Package[],
-  applicationWrapper: { app: Application | undefined },
-  wrapper: {
-    comms: DrawableClassCommunication[];
-    meshTodelete?: Application | Package | Class;
-  },
-  isCutOperation?: boolean
-) {
-  // find an ancestor that has more than 1 element
-  const foundPackage: Package | undefined = ancestorPackages.find(
-    (pckg) => pckg.subPackages.length + pckg.classes.length > 1
-  );
-  // when found delete the subpackage with the clipped object inside else check if the whole application has still more
-  // packages and if so then remove only the top package else the whole application
-  if (foundPackage) {
-    let subPackageToRemove = parentPackage;
-    const indexOfSubPackage = ancestorPackages.indexOf(foundPackage) - 1;
-    if (indexOfSubPackage >= 0)
-      subPackageToRemove = ancestorPackages[indexOfSubPackage];
-    wrapper.meshTodelete = subPackageToRemove;
-    if (isCutOperation) {
-      foundPackage.subPackages = foundPackage.subPackages.filter(
-        (pckg) => pckg.id != subPackageToRemove.id
-      );
-    }
-    return false;
-  } else {
-    const topPackage = ancestorPackages[ancestorPackages.length - 1];
-    applicationWrapper.app = getApplicationFromPackage(
-      landscapeStructure,
-      ancestorPackages[ancestorPackages.length - 1].id
-    );
-    return hasAppEnoughPackages(applicationWrapper, topPackage, wrapper);
+    deletedComms?: DrawableClassCommunication[];
   }
-  return true;
-}
-
-function hasAppEnoughPackages(
-  applicationWrapper: { app: Application | undefined },
-  topPackage: Package,
-  wrapper: {
-    comms: DrawableClassCommunication[];
-    meshTodelete?: Application | Package | Class;
-  },
-  isCutOperation?: boolean
 ) {
-  // if the application has enough packages, then remove only the package else mark the app for delete
-  if (applicationWrapper.app) {
-    if (applicationWrapper.app.packages.length > 1) {
-      wrapper.meshTodelete = topPackage;
-      if (isCutOperation) {
-        applicationWrapper.app.packages =
-          applicationWrapper.app.packages.filter(
-            (pckg: Package) => pckg.id != topPackage.id
-          );
-      }
-      return false;
-    }
-  }
-  return true;
-}
-
-function handleTopPackageAncestor(
-  landscapeStructure: StructureLandscapeData,
-  parentPackage: Package,
-  ancestorPackages: Package[],
-  applicationWrapper: { app: Application | undefined },
-  wrapper: {
-    comms: DrawableClassCommunication[];
-    meshTodelete?: Application | Package | Class;
-  },
-  isCutOperation?: boolean
-) {
-  const topPackage = ancestorPackages.firstObject;
-  // if top package has more than 1 element then remove the subpackage with the clipped object else check if the whole application has still more
-  // packages and if so then remove only the top package else the whole application
-  if (
-    topPackage &&
-    topPackage.subPackages.length + topPackage.classes.length > 1
-  ) {
-    wrapper.meshTodelete = parentPackage;
-    if (isCutOperation) {
-      topPackage.subPackages = topPackage.subPackages.filter(
-        (pckg) => pckg.id != parentPackage.id
-      );
-    }
-    return false;
-  } else if (
-    topPackage &&
-    topPackage?.subPackages.length + topPackage.classes.length <= 1
-  ) {
-    applicationWrapper.app = getApplicationFromPackage(
-      landscapeStructure,
-      topPackage.id
+  classesInApplication.forEach((clazz) => {
+    const commsToDelete = commsWrapper.comms.filter(
+      (comm) =>
+        comm.sourceClass.id === clazz.id || comm.targetClass.id === clazz.id
     );
-    return hasAppEnoughPackages(
-      applicationWrapper,
-      topPackage,
-      wrapper,
-      isCutOperation
-    );
-  }
-  return true;
+    commsWrapper.deletedComms?.pushObjects(commsToDelete);
+  });
 }
