@@ -124,6 +124,9 @@ export default class BrowserRendering extends Component<BrowserRenderingArgs> {
 
   controls!: MapControls;
 
+  ortographicCamera!: THREE.OrthographicCamera;
+  private frustumSize = 5;
+
   cameraControls!: CameraControls;
 
   initDone: boolean = false;
@@ -159,17 +162,8 @@ export default class BrowserRendering extends Component<BrowserRenderingArgs> {
     this.debug('Constructor called');
     // scene
     this.scene = defaultScene();
-    this.scene.background = this.configuration.landscapeColors.backgroundColor;
-
-    // camera
-    this.localUser.defaultCamera = new THREE.PerspectiveCamera(
-      75,
-      1.0,
-      0.1,
-      100
-    );
-    this.camera.position.set(5, 5, 5);
-    this.scene.add(this.localUser.defaultCamera);
+    this.scene.background =
+      this.configuration.applicationColors.backgroundColor;
 
     this.applicationRenderer.getOpenApplications().clear();
     // force graph
@@ -283,10 +277,21 @@ export default class BrowserRendering extends Component<BrowserRenderingArgs> {
     const width = outerDiv.clientWidth;
     const height = outerDiv.clientHeight;
 
-    // Update renderer and camera according to canvas size
+    const newAspectRatio = width / height;
+
+    // Update renderer and cameras according to canvas size
     this.renderer.setSize(width, height);
-    this.camera.aspect = width / height;
+    this.camera.aspect = newAspectRatio;
     this.camera.updateProjectionMatrix();
+
+    this.ortographicCamera.left = (this.frustumSize * newAspectRatio) / -2;
+    this.ortographicCamera.right = (this.frustumSize * newAspectRatio) / 2;
+    this.ortographicCamera.top = this.frustumSize / 2;
+    this.ortographicCamera.bottom = -this.frustumSize / 2;
+
+    this.ortographicCamera.userData.aspect = newAspectRatio;
+
+    this.ortographicCamera.updateProjectionMatrix();
   }
 
   // https://github.com/vasturiano/3d-force-graph/blob/master/example/custom-node-geometry/index.html
@@ -313,8 +318,42 @@ export default class BrowserRendering extends Component<BrowserRenderingArgs> {
     this.renderer.setSize(width, height);
     this.debug('Renderer set up');
 
+    const aspectRatio = width / height;
+
+    // camera
+    this.localUser.defaultCamera = new THREE.PerspectiveCamera(
+      75,
+      aspectRatio,
+      0.1,
+      100
+    );
+    this.camera.position.set(5, 5, 5);
+    this.scene.add(this.localUser.defaultCamera);
+
+    this.ortographicCamera = new THREE.OrthographicCamera(
+      -aspectRatio * this.frustumSize,
+      aspectRatio * this.frustumSize,
+      this.frustumSize,
+      -this.frustumSize,
+      0.1,
+      100
+    );
+
+    this.ortographicCamera.userData.aspect = aspectRatio;
+
+    this.ortographicCamera.position.setFromSphericalCoords(
+      10,
+      Math.PI / 3,
+      Math.PI / 4
+    );
+    this.ortographicCamera.lookAt(this.scene.position);
     // controls
-    this.cameraControls = new CameraControls(this.camera, this.canvas);
+    this.cameraControls = new CameraControls(
+      getOwner(this),
+      this.camera,
+      this.ortographicCamera,
+      this.canvas
+    );
 
     this.spectateUserService.cameraControls = this.cameraControls;
     this.graph.onFinishUpdate(() => {
@@ -334,6 +373,7 @@ export default class BrowserRendering extends Component<BrowserRenderingArgs> {
 
     this.renderingLoop = new RenderingLoop(getOwner(this), {
       camera: this.camera,
+      orthographicCamera: this.ortographicCamera,
       scene: this.scene,
       renderer: this.renderer,
       updatables: this.updatables,
@@ -510,6 +550,12 @@ export default class BrowserRendering extends Component<BrowserRenderingArgs> {
       this.popupHandler.clearPopups();
     } else {
       this.popupHandler.removePopup(entityId);
+    }
+
+    // remove potential toggle effect
+    const mesh = this.applicationRenderer.getMeshById(entityId);
+    if (mesh?.isHovered) {
+      mesh.resetHoverEffect();
     }
   }
 
