@@ -76,39 +76,62 @@ export default class EditMesh extends Component<Args> {
   }
 
   @action
+  duplicateApp() {
+    this.landscapeRestructure.duplicateApp(this.args.entity);
+  }
+
+  @action
   addPackage() {
     if (this.isEntityApplication)
-      this.landscapeRestructure.addPackageFromPopup(this.args.entity);
+      this.landscapeRestructure.addPackage(this.args.entity);
     else if (this.isEntityPackage)
-      this.landscapeRestructure.addSubPackageFromPopup(this.args.entity);
+      this.landscapeRestructure.addSubPackage(this.args.entity);
   }
 
   @action
   addClass() {
     if (this.isEntityPackage)
-      this.landscapeRestructure.addClassFromPopup(this.args.entity);
+      this.landscapeRestructure.addClass(this.args.entity);
   }
 
   @action
   deleteMesh() {
     if (this.isEntityApplication)
-      this.landscapeRestructure.deleteAppFromPopup(this.args.entity);
+      this.landscapeRestructure.deleteApp(this.args.entity);
     else if (this.isEntityPackage)
-      this.landscapeRestructure.deletePackageFromPopup(this.args.entity);
+      this.landscapeRestructure.deletePackage(this.args.entity);
     else if (this.isEntityClass)
-      this.landscapeRestructure.deleteClassFromPopup(this.args.entity);
+      this.landscapeRestructure.deleteClass(this.args.entity);
   }
 
   @action
   cutMesh() {
     if (this.isEntityPackage)
-      this.landscapeRestructure.cutPackageFromPopup(this.args.entity);
+      this.landscapeRestructure.cutPackage(this.args.entity);
     else if (this.isEntityClass)
-      this.landscapeRestructure.cutClassFromPopup(this.args.entity);
+      this.landscapeRestructure.cutClass(this.args.entity);
+  }
+
+  @action
+  setCommunicationSource() {
+    this.landscapeRestructure.setCommunicationSourceClass(this.args.entity);
+  }
+
+  @action
+  setCommunicationTarget() {
+    this.landscapeRestructure.setCommunicationTargetClass(this.args.entity);
   }
 
   @action
   insertMesh() {
-    this.landscapeRestructure.insertPackageOrClassFromPopup(this.args.entity);
+    this.landscapeRestructure.movePackageOrClass(this.args.entity);
+  }
+
+  @action
+  pasteMesh() {
+    if (isPackage(this.landscapeRestructure.clippedMesh))
+      this.landscapeRestructure.pastePackage(this.args.entity);
+    else if (isClass(this.landscapeRestructure.clippedMesh))
+      this.landscapeRestructure.pasteClass(this.args.entity);
   }
 }
