@@ -55,6 +55,7 @@ export default class LandscapeDataWatcherModifier extends Modifier<Args> {
 
   @service('landscape-restructure')
   landscapeRestructure!: LandscapeRestructure;
+
   @service('ide-websocket-facade')
   ideWebsocketFacade!: IdeWebsocketFacade;
 
@@ -98,8 +99,14 @@ export default class LandscapeDataWatcherModifier extends Modifier<Args> {
       this.structureLandscapeData,
       this.dynamicLandscapeData,
       this.landscapeRestructure.restructureMode,
-      this.landscapeRestructure.classCommunication
+      this.landscapeRestructure.createdClassCommunication,
+      this.landscapeRestructure.copiedClassCommunications,
+      this.landscapeRestructure.updatedClassCommunications,
+      this.landscapeRestructure.completelyDeletedClassCommunications
     );
+
+    this.landscapeRestructure.allClassCommunications =
+      drawableClassCommunications;
 
     // Use the updated landscape data to calculate application metrics.
     // This is done for all applications to have accurate heatmap data.
@@ -161,6 +168,9 @@ export default class LandscapeDataWatcherModifier extends Modifier<Args> {
       }
     }
 
+    // Apply restructure textures in restructure mode
+    this.landscapeRestructure.applyTextureMappings();
+
     const interAppCommunications = drawableClassCommunications.filter(
       (x) => x.sourceApp !== x.targetApp
     );
@@ -198,6 +208,7 @@ export default class LandscapeDataWatcherModifier extends Modifier<Args> {
           this.applicationRenderer.removeApplicationLocallyById(applicationId);
         }
       }
+      this.highlightingService.updateHighlighting();
     }
     this.graph.graphData(gData);
 
@@ -214,6 +225,9 @@ export default class LandscapeDataWatcherModifier extends Modifier<Args> {
       cls.push(tempCL);
     });
     this.ideWebsocketFacade.refreshVizData(cls);
+
+    // apply new color for restructured communications in restructure mode
+    this.landscapeRestructure.applyColorMappings();
   });
 
   updateApplicationData = task(
