@@ -1,14 +1,9 @@
 import Service, { inject as service } from '@ember/service';
 import * as THREE from 'three';
 import { tracked } from '@glimmer/tracking';
-import {
-  ApplicationColorSettingId,
-  LandscapeColorSettingId,
-} from 'explorviz-frontend/utils/settings/settings-schemas';
+import { ApplicationColorSettingId } from 'explorviz-frontend/utils/settings/settings-schemas';
 import { Position2D } from 'explorviz-frontend/modifiers/interaction-modifier';
 import UserSettings from './user-settings';
-
-export type LandscapeColors = Record<LandscapeColorSettingId, THREE.Color>;
 
 export type ApplicationColors = Record<ApplicationColorSettingId, THREE.Color>;
 
@@ -21,15 +16,6 @@ export type ApplicationColors = Record<ApplicationColorSettingId, THREE.Color>;
 export default class Configuration extends Service {
   @service('user-settings')
   userSettings!: UserSettings;
-
-  /**
-   * Colors for landscape visualization
-   *
-   * @property landscapeColors
-   * @type LandscapeColors
-   */
-  @tracked
-  landscapeColors!: LandscapeColors;
 
   /**
    * Colors for application visualization
@@ -65,22 +51,7 @@ export default class Configuration extends Service {
   constructor() {
     super(...arguments);
 
-    const { landscapeSettings, applicationSettings } = this.userSettings;
-
-    this.landscapeColors = {
-      nodeColor: new THREE.Color(landscapeSettings.nodeColor.value),
-      applicationColor: new THREE.Color(
-        landscapeSettings.applicationColor.value
-      ),
-      communicationColor: new THREE.Color(
-        landscapeSettings.communicationColor.value
-      ),
-      nodeTextColor: new THREE.Color(landscapeSettings.nodeTextColor.value),
-      applicationTextColor: new THREE.Color(
-        landscapeSettings.applicationTextColor.value
-      ),
-      backgroundColor: new THREE.Color(landscapeSettings.backgroundColor.value),
-    };
+    const { applicationSettings } = this.userSettings;
 
     this.applicationColors = {
       foundationColor: new THREE.Color(
