@@ -13,6 +13,7 @@ import CameraControls from './camera-controls';
 import { removeHighlighting } from './highlighting';
 import VrMessageSender from 'virtual-reality/services/vr-message-sender';
 import FoundationMesh from 'explorviz-frontend/view-objects/3d/application/foundation-mesh';
+import gsap from 'gsap';
 
 /**
  * Given a package or class, returns a list of all ancestor components.
@@ -59,12 +60,15 @@ export function openComponentMesh(
     return;
   }
 
-  mesh.height = 1.5;
+  gsap.to(mesh, {
+    duration: 0.25,
+    height: 1.5,
+  });
 
-  // Reset y coordinate
-  mesh.position.y -= mesh.layout.height / 2;
-  // Set y coordinate according to open component height
-  mesh.position.y += 1.5 / 2;
+  gsap.to(mesh.position, {
+    duration: 0.25,
+    y: mesh.position.y - mesh.layout.height / 2 + 0.75,
+  });
 
   mesh.opened = true;
   mesh.visible = true;
@@ -104,12 +108,15 @@ export function closeComponentMesh(
     return;
   }
 
-  mesh.height = mesh.layout.height;
+  gsap.to(mesh, {
+    duration: 0.5,
+    height: mesh.layout.height,
+  });
 
-  // Reset y coordinate
-  mesh.position.y -= 1.5 / 2;
-  // Set y coordinate according to closed component height
-  mesh.position.y += mesh.layout.height / 2;
+  gsap.to(mesh.position, {
+    duration: 0.5,
+    y: mesh.position.y - 0.75 + mesh.layout.height / 2,
+  });
 
   mesh.opened = false;
   Labeler.positionBoxLabel(mesh);
@@ -260,7 +267,14 @@ export function restoreComponentState(
     const componentMesh = applicationObject3D.getBoxMeshbyModelId(componentId);
 
     if (componentMesh) {
-      componentMesh.turnTransparent(opacity);
+      if (
+        componentMesh instanceof ClazzMesh &&
+        componentMesh.dataModel.id.includes('new')
+      ) {
+        // Without this, a new created class will be transparent
+      } else {
+        componentMesh.turnTransparent(opacity);
+      }
     }
   });
 }
