@@ -1,4 +1,8 @@
-import { Class } from './landscape-schemes/structure-data';
+import { getAllClassesInApplication } from './application-helpers';
+import {
+  Class,
+  StructureLandscapeData,
+} from './landscape-schemes/structure-data';
 import { getAncestorPackages } from './package-helpers';
 
 export function getClassAncestorPackages(clss: Class) {
@@ -11,4 +15,22 @@ export function getClassMethodHashCodes(clss: Class) {
 
 export function getClassMethodByName(clss: Class, methodName: string) {
   return clss.methods.find((method) => method.name === methodName);
+}
+
+export function getClassById(
+  landscapeData: StructureLandscapeData,
+  clssId: string
+) {
+  let clazz: Class | undefined;
+
+  if (landscapeData) {
+    for (const node of landscapeData.nodes) {
+      for (const app of node.applications) {
+        const allClasses = getAllClassesInApplication(app);
+        clazz = allClasses.find((c) => c.id === clssId);
+        if (clazz) return clazz;
+      }
+    }
+  }
+  return clazz;
 }
