@@ -3,12 +3,16 @@ import FloorMesh from 'virtual-reality/utils/view-objects/vr/floor-mesh';
 
 const FLOOR_SIZE = 1000;
 
+const PI = Math.PI;
+
 export function light(): THREE.AmbientLight {
-  return new THREE.AmbientLight(new THREE.Color(0.65, 0.65, 0.65));
+  return new THREE.AmbientLight(
+    new THREE.Color(0.65 * PI, 0.65 * PI, 0.65 * PI)
+  );
 }
 
 export function directionalLight(): THREE.DirectionalLight {
-  const light = new THREE.DirectionalLight(0xffffff, 0.55);
+  const light = new THREE.DirectionalLight(0xffffff, 0.55 * PI);
   light.name = 'DirectionalLight';
   light.position.set(-5, 5, 5);
   light.castShadow = true;
