@@ -1,8 +1,5 @@
-import {
-  defaultApplicationColors,
-  defaultLandscapeColors,
-} from './color-schemes';
-import { ApplicationSettings, LandscapeSettings } from './settings-schemas';
+import { defaultApplicationColors } from './color-schemes';
+import { ApplicationSettings } from './settings-schemas';
 
 export const defaultApplicationSettings: ApplicationSettings = {
   // Color Settings
@@ -151,6 +148,7 @@ export const defaultApplicationSettings: ApplicationSettings = {
       'If greater 0.0, communication lines are rendered arc-shaped (Straight lines: 0.0)',
     isRangeSetting: true,
   },
+  // Popup Settings
   enableCustomPopupPosition: {
     value: true,
     orderNumber: 1,
@@ -158,6 +156,24 @@ export const defaultApplicationSettings: ApplicationSettings = {
     displayName: 'Enable Custom Popup Positioning',
     description:
       'If enabled, popups can be dragged to a prefered, fixed position',
+    isFlagSetting: true,
+  },
+  // Camera Settings
+  useOrthographicCamera: {
+    value: false,
+    orderNumber: 1,
+    group: 'Camera',
+    displayName: 'Use orthographic camera instead of perspective',
+    description: 'Switch between orthographic and perspective camera',
+    isFlagSetting: true,
+  },
+  // XR Settings
+  showXRButton: {
+    value: true,
+    orderNumber: 1,
+    group: 'Extended Reality',
+    displayName: 'Show XR Button',
+    description: 'Toggle visibility of XR button',
     isFlagSetting: true,
   },
   // Debug Settings
@@ -191,69 +207,6 @@ export const defaultApplicationSettings: ApplicationSettings = {
     group: 'Debugging',
     displayName: 'Show VR in browser',
     description: 'Shows the VR room in the browser after joining',
-    isFlagSetting: true,
-  },
-};
-
-export const defaultLanscapeSettings: LandscapeSettings = {
-  // Color Settings
-  nodeColor: {
-    value: defaultLandscapeColors.nodeColor,
-    orderNumber: 1,
-    group: 'Colors',
-    displayName: 'Node',
-    isColorSetting: true,
-  },
-  applicationColor: {
-    value: defaultLandscapeColors.applicationColor,
-    orderNumber: 2,
-    group: 'Colors',
-    displayName: 'Application',
-    isColorSetting: true,
-  },
-  communicationColor: {
-    value: defaultLandscapeColors.communicationColor,
-    orderNumber: 3,
-    group: 'Colors',
-    displayName: 'Communication',
-    isColorSetting: true,
-  },
-  nodeTextColor: {
-    value: defaultLandscapeColors.nodeTextColor,
-    orderNumber: 4,
-    group: 'Colors',
-    displayName: 'Node Label',
-    isColorSetting: true,
-  },
-  applicationTextColor: {
-    value: defaultLandscapeColors.applicationTextColor,
-    orderNumber: 5,
-    group: 'Colors',
-    displayName: 'Application Label',
-    isColorSetting: true,
-  },
-  backgroundColor: {
-    value: defaultLandscapeColors.backgroundColor,
-    orderNumber: 6,
-    group: 'Colors',
-    displayName: 'Background',
-    isColorSetting: true,
-  },
-  // Hover Effect Settings
-  enableHoverEffects: {
-    value: true,
-    orderNumber: 1,
-    group: 'Hover Effects',
-    displayName: 'Enable Hover Effects',
-    description: 'Hover effect (flashing entities) for mouse cursor',
-    isFlagSetting: true,
-  },
-  showFpsCounter: {
-    value: false,
-    orderNumber: 1,
-    group: 'Debugging',
-    displayName: 'Show FPS Counter',
-    description: "'Frames Per Second' metrics in visualizations",
     isFlagSetting: true,
   },
 };
