@@ -4,24 +4,9 @@ export type SettingGroup =
   | 'Hover Effects'
   | 'Communication'
   | 'Popup'
+  | 'Camera'
+  | 'Extended Reality'
   | 'Debugging';
-
-export type LandscapeColorSettingId =
-  | 'nodeColor'
-  | 'applicationColor'
-  | 'communicationColor'
-  | 'nodeTextColor'
-  | 'applicationTextColor'
-  | 'backgroundColor';
-
-export type LandscapeHoveringSettingId = 'enableHoverEffects';
-
-export type LandscapeDebugSettingId = 'showFpsCounter';
-
-export type LandscapeSettingId =
-  | LandscapeColorSettingId
-  | LandscapeDebugSettingId
-  | LandscapeHoveringSettingId;
 
 export type ApplicationColorSettingId =
   | 'foundationColor'
@@ -47,6 +32,10 @@ export type ApplicationCommunicationSettingId =
   | 'commArrowSize'
   | 'curvyCommHeight';
 
+export type ApplicationCameraSettingId = 'useOrthographicCamera';
+
+export type ApplicationXRSettingId = 'showXRButton';
+
 export type ApplicationDebugSettingId =
   | 'showFpsCounter'
   | 'showAxesHelper'
@@ -61,22 +50,9 @@ export type ApplicationSettingId =
   | ApplicationHoveringSettingId
   | ApplicationCommunicationSettingId
   | ApplicationDebugSettingId
+  | ApplicationCameraSettingId
+  | ApplicationXRSettingId
   | ApplicationPopupSettingId;
-
-export type LandscapeColorSettings = Record<
-  LandscapeColorSettingId,
-  ColorSetting
->;
-
-export type LandscapeHoveringSettings = Record<
-  LandscapeHoveringSettingId,
-  FlagSetting
->;
-
-export type LandscapeDebugSettings = Record<
-  LandscapeDebugSettingId,
-  FlagSetting
->;
 
 export type ApplicationColorSettings = Record<
   ApplicationColorSettingId,
@@ -109,16 +85,21 @@ export type ApplicationPopupSettings = Record<
   FlagSetting
 >;
 
+export type ApplicationCameraSettings = Record<
+  ApplicationCameraSettingId,
+  FlagSetting
+>;
+
+export type ApplicationXRSettings = Record<ApplicationXRSettingId, FlagSetting>;
+
 export type ApplicationSettings = ApplicationColorSettings &
   ApplicationHighlightingSettings &
   ApplicationHoveringSettings &
   ApplicationDebugSettings &
   ApplicationPopupSettings &
+  ApplicationCameraSettings &
+  ApplicationXRSettings &
   ApplicationCommunicationSettings;
-
-export type LandscapeSettings = LandscapeColorSettings &
-  LandscapeHoveringSettings &
-  LandscapeDebugSettings;
 
 export interface Setting<T> {
   value: T;
