@@ -7,7 +7,20 @@ import debugLogger from 'ember-debug-logger';
 import Modifier, { ArgsFor } from 'ember-modifier';
 import { Position2D } from 'explorviz-frontend/modifiers/interaction-modifier';
 import ApplicationRenderer from 'explorviz-frontend/services/application-renderer';
+import Changelog from 'explorviz-frontend/services/changelog';
 import HighlightingService from 'explorviz-frontend/services/highlighting-service';
+import LandscapeRestructure from 'explorviz-frontend/services/landscape-restructure';
+import { DrawableClassCommunication } from 'explorviz-frontend/utils/application-rendering/class-communication-computer';
+import { BaseChangeLogEntry } from 'explorviz-frontend/utils/changelog-entry';
+import { getClassById } from 'explorviz-frontend/utils/class-helpers';
+import {
+  Application,
+  Class,
+  Package,
+  StructureLandscapeData,
+} from 'explorviz-frontend/utils/landscape-schemes/structure-data';
+import { getApplicationInLandscapeById } from 'explorviz-frontend/utils/landscape-structure-helpers';
+import { getPackageById } from 'explorviz-frontend/utils/package-helpers';
 import ClazzCommunicationMesh from 'explorviz-frontend/view-objects/3d/application/clazz-communication-mesh';
 import ComponentMesh from 'explorviz-frontend/view-objects/3d/application/component-mesh';
 import * as THREE from 'three';
@@ -21,6 +34,12 @@ import {
   APP_OPENED_EVENT,
 } from 'virtual-reality/utils/vr-message/sendable/app_opened';
 import {
+  CHANGELOG_REMOVE_ENTRY_EVENT,
+  CHANGELOG_RESTORE_ENTRIES_EVENT,
+  ChangeLogRemoveEntryMessage,
+  ChangeLogRestoreEntriesMessage,
+} from 'virtual-reality/utils/vr-message/sendable/changelog_update';
+import {
   ComponentUpdateMessage,
   COMPONENT_UPDATE_EVENT,
 } from 'virtual-reality/utils/vr-message/sendable/component_update';
@@ -32,6 +51,33 @@ import {
   MousePingUpdateMessage,
   MOUSE_PING_UPDATE_EVENT,
 } from 'virtual-reality/utils/vr-message/sendable/mouse-ping-update';
+import {
+  RESTRUCTURE_COMMUNICATION_EVENT,
+  RESTRUCTURE_COPY_AND_PASTE_CLASS_EVENT,
+  RESTRUCTURE_COPY_AND_PASTE_PACKAGE_EVENT,
+  RESTRUCTURE_CREATE_OR_DELETE_EVENT,
+  RESTRUCTURE_CUT_AND_INSERT_EVENT,
+  RESTRUCTURE_DELETE_COMMUNICATION_EVENT,
+  RESTRUCTURE_DUPLICATE_APP,
+  RESTRUCTURE_MODE_UPDATE_EVENT,
+  RESTRUCTURE_RENAME_OPERATION_EVENT,
+  RESTRUCTURE_RESTORE_APP_EVENT,
+  RESTRUCTURE_RESTORE_CLASS_EVENT,
+  RESTRUCTURE_RESTORE_PACKAGE_EVENT,
+  RESTRUCTURE_UPDATE_EVENT,
+  RestructureCommunicationMessage,
+  RestructureCopyAndPasteClassMessage,
+  RestructureCopyAndPastePackageMessage,
+  RestructureCreateOrDeleteMessage,
+  RestructureCutAndInsertMessage,
+  RestructureDeleteCommunicationMessage,
+  RestructureDuplicateAppMessage,
+  RestructureRenameOperationMessage,
+  RestructureRestoreAppMessage,
+  RestructureRestoreClassMessage,
+  RestructureRestorePackageMessage,
+  RestructureUpdateMessage,
+} from 'virtual-reality/utils/vr-message/sendable/restructure_update';
 
 interface IModifierArgs {
   positional: [];
@@ -68,6 +114,78 @@ export default class CollaborativeModifierModifier extends Modifier<IModifierArg
       this,
       this.onHighlightingUpdate
     );
+    this.webSocket.on(
+      RESTRUCTURE_MODE_UPDATE_EVENT,
+      this,
+      this.onRestructureModeUpdate
+    );
+    this.webSocket.on(RESTRUCTURE_UPDATE_EVENT, this, this.onRestructureUpdate);
+    this.webSocket.on(
+      RESTRUCTURE_CREATE_OR_DELETE_EVENT,
+      this,
+      this.onRestructureCreateOrDelete
+    );
+    this.webSocket.on(
+      RESTRUCTURE_COPY_AND_PASTE_PACKAGE_EVENT,
+      this,
+      this.onRestructureCopyAndPastePackage
+    );
+    this.webSocket.on(
+      RESTRUCTURE_COPY_AND_PASTE_CLASS_EVENT,
+      this,
+      this.onRestructureCopyAndPasteClass
+    );
+    this.webSocket.on(
+      RESTRUCTURE_CUT_AND_INSERT_EVENT,
+      this,
+      this.onRestructureCutAndInsert
+    );
+    this.webSocket.on(
+      RESTRUCTURE_COMMUNICATION_EVENT,
+      this,
+      this.onRestructureCommunication
+    );
+    this.webSocket.on(
+      RESTRUCTURE_DELETE_COMMUNICATION_EVENT,
+      this,
+      this.onRestructureDeleteCommunication
+    );
+    this.webSocket.on(
+      RESTRUCTURE_RENAME_OPERATION_EVENT,
+      this,
+      this.onRestructureRenameOperationMessage
+    );
+    this.webSocket.on(
+      RESTRUCTURE_RESTORE_APP_EVENT,
+      this,
+      this.onRestructureRestoreApp
+    );
+    this.webSocket.on(
+      RESTRUCTURE_RESTORE_PACKAGE_EVENT,
+      this,
+      this.onRestructureRestorePackage
+    );
+    this.webSocket.on(
+      RESTRUCTURE_RESTORE_CLASS_EVENT,
+      this,
+      this.onRestructureRestoreClass
+    );
+    this.webSocket.on(
+      CHANGELOG_REMOVE_ENTRY_EVENT,
+      this,
+      this.onChangeLogRemoveEntry
+    );
+    this.webSocket.on(
+      CHANGELOG_RESTORE_ENTRIES_EVENT,
+      this,
+      this.onChangeLogRestoreEntriesMessage
+    );
+    this.webSocket.on(
+      RESTRUCTURE_DUPLICATE_APP,
+      this,
+      this.onRestructureDuplicateApp
+    );
+
     registerDestructor(this, this.cleanup);
   }
 
@@ -84,6 +202,81 @@ export default class CollaborativeModifierModifier extends Modifier<IModifierArg
       this,
       this.onHighlightingUpdate
     );
+    this.webSocket.off(
+      RESTRUCTURE_MODE_UPDATE_EVENT,
+      this,
+      this.onRestructureModeUpdate
+    );
+    this.webSocket.off(
+      RESTRUCTURE_UPDATE_EVENT,
+      this,
+      this.onRestructureUpdate
+    );
+    this.webSocket.off(
+      RESTRUCTURE_CREATE_OR_DELETE_EVENT,
+      this,
+      this.onRestructureCreateOrDelete
+    );
+    this.webSocket.off(
+      RESTRUCTURE_COPY_AND_PASTE_PACKAGE_EVENT,
+      this,
+      this.onRestructureCopyAndPastePackage
+    );
+    this.webSocket.off(
+      RESTRUCTURE_COPY_AND_PASTE_CLASS_EVENT,
+      this,
+      this.onRestructureCopyAndPasteClass
+    );
+    this.webSocket.off(
+      RESTRUCTURE_CUT_AND_INSERT_EVENT,
+      this,
+      this.onRestructureCutAndInsert
+    );
+    this.webSocket.off(
+      RESTRUCTURE_COMMUNICATION_EVENT,
+      this,
+      this.onRestructureCommunication
+    );
+    this.webSocket.off(
+      RESTRUCTURE_DELETE_COMMUNICATION_EVENT,
+      this,
+      this.onRestructureDeleteCommunication
+    );
+    this.webSocket.off(
+      RESTRUCTURE_RENAME_OPERATION_EVENT,
+      this,
+      this.onRestructureRenameOperationMessage
+    );
+    this.webSocket.off(
+      RESTRUCTURE_RESTORE_APP_EVENT,
+      this,
+      this.onRestructureRestoreApp
+    );
+    this.webSocket.off(
+      RESTRUCTURE_RESTORE_PACKAGE_EVENT,
+      this,
+      this.onRestructureRestorePackage
+    );
+    this.webSocket.off(
+      RESTRUCTURE_RESTORE_CLASS_EVENT,
+      this,
+      this.onRestructureRestoreClass
+    );
+    this.webSocket.off(
+      CHANGELOG_REMOVE_ENTRY_EVENT,
+      this,
+      this.onChangeLogRemoveEntry
+    );
+    this.webSocket.off(
+      CHANGELOG_RESTORE_ENTRIES_EVENT,
+      this,
+      this.onChangeLogRestoreEntriesMessage
+    );
+    this.webSocket.off(
+      RESTRUCTURE_DUPLICATE_APP,
+      this,
+      this.onRestructureDuplicateApp
+    );
   }
 
   debug = debugLogger('CollaborativeModifier');
@@ -99,6 +292,12 @@ export default class CollaborativeModifierModifier extends Modifier<IModifierArg
 
   @service('highlighting-service')
   private highlightingService!: HighlightingService;
+
+  @service('landscape-restructure')
+  landscapeRestructure!: LandscapeRestructure;
+
+  @service('changelog')
+  changeLog!: Changelog;
 
   @service('local-user')
   private localUser!: LocalUser;
@@ -167,7 +366,6 @@ export default class CollaborativeModifierModifier extends Modifier<IModifierArg
   }: ForwardedMessage<HighlightingUpdateMessage>): void {
     const user = this.collaborationSession.lookupRemoteUserById(userId);
     if (!user) return;
-
     const application = this.applicationRenderer.getApplicationById(appId);
     if (!application) {
       // extern communication link
@@ -187,6 +385,256 @@ export default class CollaborativeModifierModifier extends Modifier<IModifierArg
       multiSelected,
       false // whenever we receive messages we don't want to resend them
     );
+  }
+
+  onRestructureModeUpdate(): void {
+    this.landscapeRestructure.toggleRestructureModeLocally();
+  }
+
+  onRestructureUpdate({
+    originalMessage: { entityType, entityId, newName, appId, undo },
+  }: ForwardedMessage<RestructureUpdateMessage>): void {
+    switch (entityType) {
+      case 'APP':
+        this.landscapeRestructure.renameApplication(
+          newName,
+          entityId,
+          true,
+          undo
+        );
+        break;
+      case 'PACKAGE':
+        this.landscapeRestructure.renamePackage(newName, entityId, true, undo);
+        break;
+      case 'SUBPACKAGE':
+        this.landscapeRestructure.renameSubPackage(
+          newName,
+          entityId,
+          true,
+          undo
+        );
+        break;
+      case 'CLAZZ':
+        this.landscapeRestructure.renameClass(
+          newName,
+          entityId,
+          appId as string,
+          true,
+          undo
+        );
+    }
+  }
+
+  onRestructureCreateOrDelete({
+    originalMessage: { action, entityType, name, language, entityId, undo },
+  }: ForwardedMessage<RestructureCreateOrDeleteMessage>): void {
+    if (action === 'CREATE') {
+      switch (entityType) {
+        case 'APP':
+          this.landscapeRestructure.addApplication(
+            name as string,
+            language as string,
+            true
+          );
+          break;
+        case 'PACKAGE':
+          this.landscapeRestructure.addCollaborativePackage(entityId as string);
+          break;
+        case 'SUBPACKAGE':
+          this.landscapeRestructure.addCollaborativeSubPackage(
+            entityId as string
+          );
+          break;
+        case 'CLAZZ':
+          this.landscapeRestructure.addCollaborativeClass(entityId as string);
+          break;
+      }
+    } else {
+      switch (entityType) {
+        case 'APP':
+          this.landscapeRestructure.deleteCollaborativeApplication(
+            entityId as string,
+            undo as boolean
+          );
+          break;
+        case 'PACKAGE':
+          this.landscapeRestructure.deleteCollaborativePackage(
+            entityId as string,
+            undo as boolean
+          );
+          break;
+        case 'CLAZZ':
+          this.landscapeRestructure.deleteCollaborativeClass(
+            entityId as string,
+            undo as boolean
+          );
+          break;
+      }
+    }
+  }
+
+  onRestructureDuplicateApp({
+    originalMessage: { appId },
+  }: ForwardedMessage<RestructureDuplicateAppMessage>): void {
+    const app = getApplicationInLandscapeById(
+      this.landscapeRestructure.landscapeData
+        ?.structureLandscapeData as StructureLandscapeData,
+      appId
+    );
+    this.landscapeRestructure.duplicateApp(app as Application, true);
+  }
+
+  onRestructureCopyAndPastePackage({
+    originalMessage: { destinationEntity, destinationId, clippedEntityId },
+  }: ForwardedMessage<RestructureCopyAndPastePackageMessage>): void {
+    this.landscapeRestructure.pasteCollaborativePackage(
+      destinationEntity,
+      destinationId,
+      clippedEntityId
+    );
+  }
+
+  onRestructureCopyAndPasteClass({
+    originalMessage: { destinationId, clippedEntityId },
+  }: ForwardedMessage<RestructureCopyAndPasteClassMessage>): void {
+    this.landscapeRestructure.pasteCollaborativeClass(
+      destinationId,
+      clippedEntityId
+    );
+  }
+
+  onRestructureCutAndInsert({
+    originalMessage: {
+      destinationEntity,
+      destinationId,
+      clippedEntity,
+      clippedEntityId,
+    },
+  }: ForwardedMessage<RestructureCutAndInsertMessage>): void {
+    this.landscapeRestructure.insertCollaborativePackagerOrClass(
+      destinationEntity,
+      destinationId,
+      clippedEntity,
+      clippedEntityId
+    );
+  }
+
+  onRestructureCommunication({
+    originalMessage: { sourceClassId, targetClassId, methodName },
+  }: ForwardedMessage<RestructureCommunicationMessage>): void {
+    this.landscapeRestructure.addCollaborativeCommunication(
+      sourceClassId,
+      targetClassId,
+      methodName
+    );
+  }
+
+  onRestructureDeleteCommunication({
+    originalMessage: { undo, commId },
+  }: ForwardedMessage<RestructureDeleteCommunicationMessage>): void {
+    const comm = this.landscapeRestructure.allClassCommunications.find(
+      (comm) => comm.id === commId
+    );
+    this.landscapeRestructure.deleteCommunication(
+      comm as DrawableClassCommunication,
+      undo,
+      true
+    );
+  }
+
+  onRestructureRenameOperationMessage({
+    originalMessage: { commId, newName, undo },
+  }: ForwardedMessage<RestructureRenameOperationMessage>): void {
+    const comm = this.landscapeRestructure.allClassCommunications.find(
+      (comm) => comm.id === commId
+    );
+
+    this.landscapeRestructure.renameOperation(
+      comm as DrawableClassCommunication,
+      newName,
+      true,
+      undo
+    );
+  }
+
+  onRestructureRestoreApp({
+    originalMessage: { appId, undoCutOperation },
+  }: ForwardedMessage<RestructureRestoreAppMessage>): void {
+    const landscapeData =
+      this.landscapeRestructure.landscapeData?.structureLandscapeData;
+    const app = getApplicationInLandscapeById(
+      landscapeData as StructureLandscapeData,
+      appId
+    );
+    this.landscapeRestructure.restoreApplication(
+      app as Application,
+      undoCutOperation,
+      true
+    );
+  }
+  onRestructureRestorePackage({
+    originalMessage: { pckgId, undoCutOperation },
+  }: ForwardedMessage<RestructureRestorePackageMessage>): void {
+    const landscapeData =
+      this.landscapeRestructure.landscapeData?.structureLandscapeData;
+    const pckg = getPackageById(
+      landscapeData as StructureLandscapeData,
+      pckgId
+    );
+    this.landscapeRestructure.restorePackage(
+      pckg as Package,
+      undoCutOperation,
+      true
+    );
+  }
+
+  onRestructureRestoreClass({
+    originalMessage: { appId, clazzId, undoCutOperation },
+  }: ForwardedMessage<RestructureRestoreClassMessage>): void {
+    const landscapeData =
+      this.landscapeRestructure.landscapeData?.structureLandscapeData;
+    const app = getApplicationInLandscapeById(
+      landscapeData as StructureLandscapeData,
+      appId
+    );
+    const clazz = getClassById(
+      landscapeData as StructureLandscapeData,
+      clazzId
+    );
+    this.landscapeRestructure.restoreClass(
+      app as Application,
+      clazz as Class,
+      undoCutOperation,
+      true
+    );
+  }
+
+  onChangeLogRemoveEntry({
+    originalMessage: { entryIds },
+  }: ForwardedMessage<ChangeLogRemoveEntryMessage>): void {
+    if (entryIds.length > 1) {
+      const entries: BaseChangeLogEntry[] = [];
+      entryIds.forEach((id) => {
+        const foundEntry = this.changeLog.changeLogEntries.find(
+          (entry) => entry.id === id
+        );
+        entries.push(foundEntry as BaseChangeLogEntry);
+      });
+
+      this.changeLog.removeEntries(entries, true);
+    } else {
+      const foundEntry = this.changeLog.changeLogEntries.find(
+        (entry) => entry.id === entryIds[0]
+      );
+
+      this.changeLog.removeEntry(foundEntry as BaseChangeLogEntry, true);
+    }
+  }
+
+  onChangeLogRestoreEntriesMessage({
+    originalMessage: { key },
+  }: ForwardedMessage<ChangeLogRestoreEntriesMessage>): void {
+    this.changeLog.restoreDeletedEntries(key, true);
   }
 
   onMousePingUpdate({
